+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   than rounding to the nearest 1000 kilobytes (#28442).
 - Continue running with defaults when the config argument is provided
   but not found (#28500).
+- Make sbws round to 2 significant figures by default. This implements part
+  of proposal 276 (#28451).
 
 ## [1.0.2] - 2018-11-10